% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backcalc-package.R
\docType{package}
\name{backcalc}
\alias{backcalc}
\alias{backcalc-package}
\title{backcalc: Reconstruct Missing Inferential Statistics}
\description{
This package provides functions to reconstruct key inferential statistics such as standard errors,
confidence intervals, test statistics, and p-values from partial or incomplete summary information
commonly reported in scientific literature.
}
\details{
\code{backcalc} is an R package designed to help researchers and students reconstruct key inferential statistics —-
such as standard errors, confidence intervals, test statistics, and p-values —-
from partial or incomplete summary data commonly found in scientific literature.
Supporting various data types and designs, including single- and two-sample (independent or paired) tests
with z or t distributions, it offers tailored functions for means, medians,
correlations, regression coefficients, proportions, and ratio measures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cwendorf/backcalc}
  \item Report bugs at \url{https://github.com/cwendorf/backcalc/issues}
}

}
\author{
\strong{Maintainer}: Craig Wendorf
\email{cwendorf@uwsp.edu}
}
