% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backcalc_print.R
\name{knit_print.backcalc}
\alias{knit_print.backcalc}
\title{Knit print method for backcalc objects
Ensures Notes and Approximations print in Rmd}
\usage{
knit_print.backcalc(x, ...)
}
\description{
Knit print method for backcalc objects
Ensures Notes and Approximations print in Rmd
}
