% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backcalc-package.R
\docType{package}
\name{backcalc}
\alias{backcalc}
\alias{backcalc-package}
\title{backcalc: Reconstruct Missing Inferential Statistics}
\description{
Tools to reconstruct inferential statistics -- like standard errors, p-values, and confidence intervals --
from incomplete summary data reported in scientific research.
}
\details{
This package provides a set of functions to reconstruct key inferential statistics
such as standard errors, confidence intervals, test statistics, and p-values
from partial or incomplete summary information commonly reported in scientific
literature. The functions support a variety of data types and designs, including
single-sample, two-sample (independent or paired), one- and two-sided tests,
and accommodate both z and t distributions.

Specifically, the package offers tailored functions for:
\itemize{
\item Means (\code{backcalc_means})
\item Regression coefficients (\code{backcalc_coeffs})
\item Correlations (\code{backcalc_corrs})
\item Ratio measures (\code{backcalc_ratios})
\item Proportions (\code{backcalc_props})
}

These tools facilitate meta-analysts, researchers, and students in recovering missing
inferential details from published results to enable accurate synthesis, reporting,
or further analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cwendorf/backcalc}
  \item Report bugs at \url{https://github.com/cwendorf/backcalc/issues}
}

}
\author{
\strong{Maintainer}: Craig Wendorf
\email{cwendorf@uwsp.edu}
}
